TST: included probe_classes module
